Refactor MusicBrainz provider into multi-file package
Splits the monolithic musicbrainz/__init__.py (718 lines) into a proper
multi-file package for better maintainability and easier extension.

New structure:
- constants.py: Shared config (LUCENE_SPECIAL, URL_RELATION_TYPE_MAPPING, etc.)
- models.py: All MusicBrainz dataclasses and replace_hyphens helper
- api_client.py: HTTP client with 10 req/10s throttling and 30-day caching
- provider.py: MusicbrainzProvider class with all metadata functionality
- __init__.py: Re-exports public API (setup, get_config_entries, MusicbrainzProvider)

Updated imports in:
- controllers/metadata/radio.py: Import MusicBrainzReleaseGroup from .models
- providers/wikipedia/__init__.py: Import MusicBrainzRelation from .models

No functional changes - all existing behavior preserved including:
- URL relations support (Wikipedia, social links, etc.)
- ISRC lookup with fallback for merged/redirected MBIDs
- Artist metadata enrichment
- Release group search by track name
